Refactored ThemeContext for seperation of concerns
