+++ src/contexts/ThemeContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { ThemeContextType } from '../types/theme';
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
