+++ src/main.tsx
@@ -1,11 +1,11 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
+import React from 'react';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { ThemeProvider } from './contexts/ThemeContext';
-import App from './App.tsx'
-import './i18n'
+import ThemeProvider from './providers/ThemeProvider';
+import App from './App.tsx';
+import './i18n';
 import 'react-loading-skeleton/dist/skeleton.css';
-import './styles/index.scss'
+import './styles/index.scss';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
@@ -14,5 +14,5 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
         <App />
       </ThemeProvider>
     </Router>
-  </React.StrictMode>,
-)
+  </React.StrictMode>
+);
