+++ src/types/theme.ts
@@ -1,0 +1,6 @@
+export type Theme = 'light' | 'dark';
+
+export interface ThemeContextType {
+  theme: Theme;
+  toggleTheme: () => void;
+}
